FunctionApp2/requirements.txt
pyodbc
requests
azure-functions
